Refactor project structure and release bot
