--- .github/release-drafter.yml
@@ -1,4 +1,0 @@
-template: |
-  ## What’s Changed
-
-  $CHANGES
