--- .github/workflows/auth.yml
@@ -1,20 +1,0 @@
-name: auth
-
-on: push
-
-jobs:
-  login:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: 12
-      - run: npm ci
-      - run: npm run build
-      - run: node auth.js "US"
-      - run: node auth.js "EU"
-    env:
-      HYUNDAI_USER: ${{secrets.HYUNDAI_USER}}
-      HYUNDAI_PASS: ${{secrets.HYUNDAI_PASS}}
-      HYUNDAI_PIN: ${{secrets.HYUNDAI_PIN}}
