+++ __tests__/bluelinky.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable */
 
 import got from 'got';
-import BlueLinky from '../index';
+import BlueLinky from '../src/index';
 jest.mock('got');
 
 const gotMock = got as any;
