+++ __tests__/controller.spec.ts
@@ -1,9 +1,9 @@
 /* eslint-disable */
 import got from 'got';
 
-import { AmericanController } from '../controllers/american.controller';
-import { EuropeanController } from '../controllers/european.controller';
-import { CanadianController } from '../controllers/canadian.controller';
+import { AmericanController } from '../src/controllers/american.controller';
+import { EuropeanController } from '../src/controllers/european.controller';
+import { CanadianController } from '../src/controllers/canadian.controller';
 
 jest.mock('got');
