+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "build": "rollup -c rollup.config.js",
     "dev": "rollup -w -c rollup.config.js",
-    "lint": "eslint ./lib --ext .js,.jsx,.ts,.tsx",
+    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx",
     "prepublishOnly": "npm run build",
     "test": "jest --verbose",
     "debug": "cross-env LOG_LEVEL=debug ts-node debug.ts",
