+++ rollup.config.js
@@ -1,11 +1,11 @@
 import typescript from 'rollup-plugin-typescript2';
 import resolve from 'rollup-plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
-import { terser } from "rollup-plugin-terser";
+import { terser } from 'rollup-plugin-terser';
 import pkg from './package.json';
 
 export default {
-  input: 'lib/index.ts',
+  input: 'src/index.ts',
   output: {
     format: 'cjs',
     name: 'index',
