+++ src/constants/europe.ts
