+++ src/interfaces/common.interfaces.ts
