+++ src/interfaces/european.interfaces.ts
