+++ src/vehicles/american.vehicle.ts
