+++ src/vehicles/european.vehicle.ts
